fix(android): avoid double-inclusion of vector-icons fonts
we include the react-native-vector-icons gradle script in our build.gradle,
that script already takes care of merging in the font files as needed

Including them in a fixed location in android assets is actually a
double-inclusion. This was hidden before because the vector-icons
gradle script moved them to the same spot, but with v9.1.0 of that
package it exposed the issue with a build error.

Delete them here, android still has icons, Fixes #49